PrestoGPModel superclass

This is the superclass for PrestoGP models. All other types of PrestoGP model classes (e.g. VecchiaModel-class and FullModel-class are inherited from this class. Normally users should not create objects of this class. Instead, they should use the appropriate inherited class for the type of model they are fitting.

Slots

covparams

A numeric vector containing the parameters for the Matern model.

logparams

A numeric vector containing the transformed versions of the Matern parameters (used internally for likelihood calculations).

beta

A matrix containing the regression coefficients.

lambda_1se_idx

Stores the index of the optimal tuning parameter for the glmnet model. See glmnet.

vecchia_approx

The output of the Vecchia specify function. See vecchia_specify and vecchia_Mspecify.

X_train

A matrix containing the original predictors. This will be a "super matrix" for multivariate models. See superMatrix.

Y_train

A column matrix containing the original response values.

X_tilde

The matrix of transformed predictors.

y_tilde

The column matrix containing the transformed response values.

res

numeric.

locs_train

A list containing the location coordinates. Each element of the list corresponds to a different outcome. (The list will have length 1 for univariate models.)

res

A numeric vector of the residuals.

linear_model

The glmnet model. See glmnet and cv.glmnet.

converged

Did the model fitting process converge (boolean)?

LL_Vecchia_krig

The value of the negative log likelihood function after optimization.

error

Penalized model error. See References for details.

n_eighbors

Number of neighbors to condition on for the Vecchia approximation. Ignored for full models.

min_m

Minimum permissible number of neighbors.

alpha

Parameter alpha for glmnet. See glmnet.

scaling

The indices of the scale parameters. See link{prestogp_fit}.

nscale

The number of scale parameters in the model.

apanasovich

Should the Apanasovich covariance model be used? See References.

param_sequence

Records the indices of the various Matern parameters. See create.param.sequence.

References

Apanasovich, T.V., Genton, M.G. and Sun, Y. "A valid Matérn class of cross-covariance functions for multivariate random fields with any number of components", Journal of the American Statistical Association (2012) 107(497):180-193.

Messier, K.P. and Katzfuss, M. "Scalable penalized spatiotemporal land-use regression for ground-level nitrogen dioxide", The Annals of Applied Statistics (2021) 15(2):688-710.

See also

Examples

pgp.vmodel <- new("VecchiaModel", n_neighbors = 25)
pgp.fmodel <- new("FullModel")
pgp.mmodel <- new("MultivariateVecchiaModel", n_neighbors = 25)